third_party/blink/web_tests/serial/serialPort_readable.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/mojo/public/mojom/base/unguessable_token.mojom.js"></script>
<script src="file:///gen/third_party/blink/public/mojom/serial/serial.mojom.js"></script>
<script src="resources/serial-test-utils.js"></script>
<script>

function compareArrays(actual, expected) {
  assert_equals(actual.byteLength, expected.byteLength);
  for (let i = 0; i < expected.byteLength; ++i)
    assert_equals(actual[i], expected[i], `Mismatch at position ${i}.`);
}

async function readAll(reader) {
  const chunks = [];
  while (true) {
    let { value, done } = await reader.read();
    if (done) {
      // It would be better to allocate |buffer| up front with the number of
      // of bytes expected but this is the best that can be done without a BYOB
      // reader to control the amount of data read.
      const length =
          chunks.reduce((total, chunk) => total + chunk.byteLength, 0);
      const buffer = new Uint8Array(length);
      chunks.reduce((offset, chunk) => {
        buffer.set(chunk, offset);
        return offset + chunk.byteLength;
      }, 0);
      return buffer;
    }

    chunks.push(value);
  }
}

serial_test(async (t, fake) => {
  const { port, fakePort } = await getFakeSerialPort(fake);

  assert_equals(port.readable, null);

  await port.open({ baudrate: 9600 });
  assert_true(port.readable instanceof ReadableStream);

  port.close();
  assert_equals(port.readable, null);
}, 'open() and close() set and unset SerialPort.readable');

serial_test(async (t, fake) => {
  const { port, fakePort } = await getFakeSerialPort(fake);

  assert_equals(port.readable, null);

  await port.open({ baudrate: 9600 });
  assert_true(port.readable instanceof ReadableStream);

  const reader = port.readable.getReader();
  port.close();

  const { value, done } = await reader.read();
  assert_true(done);
  assert_equals(value, undefined);
  assert_equals(port.readable, null);
}, 'SerialPort.readable closes on port close');

serial_test(async (t, fake) => {
  const { port, fakePort } = await getFakeSerialPort(fake);
  // Select a buffer size larger than the amount of data transferred.
  await port.open({ baudrate: 9600, buffersize: 64 });

  const data = new Uint8Array([1, 2, 3, 4, 5, 6, 7, 8]);
  fakePort.write(data);

  const reader = port.readable.getReader();
  let { value, done } = await reader.read();
  assert_false(done);
  compareArrays(data, value);

  port.close();

  ({ value, done } = await reader.read());
  assert_true(done);
  assert_equals(undefined, value);
}, 'Can read a small amount of data');

serial_test(async (t, fake) => {
  const { port, fakePort } = await getFakeSerialPort(fake);
  // Select a buffer size smaller than the amount of data transferred.
  await port.open({ baudrate: 9600, buffersize: 64 });

  const data = new Uint8Array(1024);  // Much larger than buffersize above.
  for (let i = 0; i < data.byteLength; ++i)
    data[i] = i & 0xff;
  fakePort.write(data);

  const reader = port.readable.getReader();
  port.close();
  const value = await readAll(reader);
  compareArrays(data, value);
}, 'Can read a large amount of data');

serial_test(async (t, fake) => {
  const { port, fakePort } = await getFakeSerialPort(fake);
  // Select a buffer size smaller than the amount of data transferred.
  await port.open({ baudrate: 9600, buffersize: 64 });

  const data = new Uint8Array([1, 2, 3, 4, 5, 6, 7, 8]);
  fakePort.write(data);
  fakePort.simulateParityError();

  let reader = port.readable.getReader();
  let { value, done } = await reader.read();
  assert_false(done);
  compareArrays(data, value);

  await promise_rejects(t, 'NetworkError', reader.read());
  assert_equals(port.readable, null);

  port.clearReadError();
  assert_true(port.readable instanceof ReadableStream);
  reader = port.readable.getReader();

  await fakePort.waitForErrorCleared();
  fakePort.write(data);

  ({ value, done } = await reader.read());
  assert_false(done);
  compareArrays(data, value);

  port.close();
  ({ value, done } = await reader.read());
  assert_true(done);
  assert_equals(undefined, value);
}, 'Parity error closes readable, can be cleared and more data received');

</script>
